requirements.txt
cffi==1.16.0
cryptography==42.0.3
paho-mqtt==2.0.0
pycparser==2.21
